backend/app.py
import os
import numpy as np
import tensorflow as tf
from flask import Flask, request, jsonify
from flask_cors import CORS
from dotenv import load_dotenv
from supabase import create_client, Client
from PIL import Image
from werkzeug.utils import secure_filename
import io
import cv2
import time
import uuid

# Load environment variables
load_dotenv()

app = Flask(__name__)
CORS(app)

# Supabase Setup
url = os.environ.get("SUPABASE_URL")
key = os.environ.get("SUPABASE_KEY")
supabase: Client = create_client(url, key) if url and key else None

# Model Setup
MODEL_PATH = os.path.join(os.path.dirname(__file__), 'models', 'VGG16_mammogram_model.h5')
model = None

def load_model():
    global model
    try:
        if os.path.exists(MODEL_PATH):
            print(f"Loading model from {MODEL_PATH}...")
            model = tf.keras.models.load_model(MODEL_PATH)
            print("Model loaded successfully.")
        else:
            print(f"Model not found at {MODEL_PATH}")
    except Exception as e:
        print(f"Error loading model: {e}")

load_model()

def preprocess_image(image_bytes):
    # Convert bytes to PIL Image
    img = Image.open(io.BytesIO(image_bytes))
    img = img.convert('RGB')
    
    # Resize to model input size (150x150 based on VGG16 variant/User Feedback)
    target_size = (150, 150)
    img = img.resize(target_size)
    
    # Convert to array and normalize
    img_array = np.array(img)
    img_array = img_array / 255.0
    img_array = np.expand_dims(img_array, axis=0)
    
    return img_array, img

@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({"status": "healthy", "model_loaded": model is not None})

@app.route('/predict', methods=['POST'])
def predict():
    if not model:
        return jsonify({"error": "Model not loaded"}), 500

    # Verify Auth
    auth_header = request.headers.get('Authorization')
    if not auth_header:
        # Check if we can proceed without auth for testing, or fail?
        # Specification says "Verification: Middleware to verify... before processing"
        # Since we need user_id for DB, we MUST have auth.
        return jsonify({"error": "Missing Authorization header"}), 401
    
    token = auth_header.split(" ")[1]
    
    try:
        user_response = supabase.auth.get_user(token)
        user_id = user_response.user.id
    except Exception as e:
        print(f"Auth error: {e}")
        return jsonify({"error": "Invalid token"}), 401

    if 'file' not in request.files:
        return jsonify({"error": "No file part"}), 400
    
    file = request.files['file']
    if file.filename == '':
        return jsonify({"error": "No selected file"}), 400

    try:
        # Read file
        file_bytes = file.read()
        filename = secure_filename(file.filename)
        
        # Preprocess
        processed_img, original_pil = preprocess_image(file_bytes)
        
        # Predict
        prediction = model.predict(processed_img)
        
        # Prediction is [[prob_benign, prob_malignant]]
        # Handle different output shapes just in case
        if prediction.shape[-1] == 2:
            prob_benign = prediction[0][0]
            prob_malignant = prediction[0][1]
            label = "Malignant" if prob_malignant > prob_benign else "Benign"
            confidence = float(prob_malignant) if label == "Malignant" else float(prob_benign)
        else:
             # Binary single neuron output
             confidence_val = float(prediction[0][0])
             label = "Malignant" if confidence_val > 0.5 else "Benign"
             confidence = confidence_val if label == "Malignant" else 1 - confidence_val

        # --- Supabase Integration ---
        
        # 1. Upload Original Image
        # Create a unique path
        unique_id = str(uuid.uuid4())
        file_ext = filename.rsplit('.', 1)[1].lower() if '.' in filename else 'jpg'
        storage_path = f"{user_id}/{unique_id}.{file_ext}"
        
        # Reset file pointer or use bytes
        try:
            res = supabase.storage.from_("mammo-scans").upload(
                path=storage_path,
                file=file_bytes,
                file_options={"content-type": file.content_type}
            )
            # Get Public URL
            public_url_res = supabase.storage.from_("mammo-scans").get_public_url(storage_path)
            # Depending on supabase version, get_public_url might return string or string inside logic
            # Usually it returns the URL string directly
            image_url = public_url_res
            
        except Exception as storage_err:
            print(f"Storage Error: {storage_err}")
            # Fallback to empty URL if storage fails, but proceed (or fail?)
            # Let's fail for now as storage is critical requirement
            # If bucket doesn't exist, this will fail.
            image_url = ""
            print("Ensure 'mammo-scans' bucket exists and is public.")

        # 2. Save to Database
        if image_url:
            db_data = {
                "user_id": user_id,
                "original_image_url": image_url,
                "prediction_label": label,
                "confidence_score": confidence,
                "annotated_image_url": image_url # For now same as original
            }
            db_res = supabase.table("scans").insert(db_data).execute()
        
        return jsonify({
            "prediction": label,
            "confidence": confidence,
            "image_url": image_url,
            "raw_output": prediction.tolist()
        })

    except Exception as e:
        print(f"Error processing: {e}")
        import traceback
        traceback.print_exc()
        return jsonify({"error": str(e)}), 500

@app.route('/scans/<scan_id>', methods=['DELETE'])
def delete_scan(scan_id):
    if not model:
         return jsonify({"error": "Model not loaded"}), 500

    # Verify Auth
    auth_header = request.headers.get('Authorization')
    if not auth_header:
        return jsonify({"error": "Missing Authorization header"}), 401
    
    token = auth_header.split(" ")[1]
    
    try:
        user_response = supabase.auth.get_user(token)
        user_id = user_response.user.id
    except Exception as e:
        print(f"Auth error: {e}")
        return jsonify({"error": "Invalid token"}), 401

    try:
        # 1. Fetch scan to get storage path
        res = supabase.table("scans").select("*").eq("id", scan_id).execute()
        
        if not res.data or len(res.data) == 0:
            return jsonify({"error": "Scan not found or access denied"}), 404
        
        scan = res.data[0]
        # Verify ownership
        if scan.get('user_id') != user_id:
             return jsonify({"error": "Unauthorized"}), 403

        # 2. Delete from Storage
        original_url = scan.get('original_image_url')
        if original_url:
            if "mammo-scans/" in original_url:
                # Extract path after mammo-scans/
                # e.g. .../mammo-scans/userid/file.jpg
                # We need userid/file.jpg
                storage_path = original_url.split("mammo-scans/")[1]
                storage_path = storage_path.split("?")[0]
                
                print(f"Deleting file: {storage_path}")
                supabase.storage.from_("mammo-scans").remove([storage_path])

        # 3. Delete from Database
        supabase.table("scans").delete().eq("id", scan_id).execute()
        
        return jsonify({"message": "Scan deleted successfully"})

    except Exception as e:
        print(f"Delete error: {e}")
        return jsonify({"error": str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)
