backend/requirements.txt
flask
flask-cors
tensorflow
numpy
opencv-python-headless
supabase
python-dotenv
werkzeug
pillow
